# Conflicts:
#	src/main/java/com/example/demo/stock/repository/OrdersRepository.java

+++ src/main/java/com/example/demo/stock/dto/res/RoundResultResponse.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Builder
 public record RoundResultResponse(
         Integer roundNumber,
-        Integer year,
+        Long year,
         List<TeamInvestmentDto> teamInvestments
 ) {
